Updating the constructor recipe.
Adding menuinst and console_shortcut packages on windows so that the
anaconda console is available from the start menu. This allows us
to open the jupyter notebook anywhere in the file system. The jupyter
notebook shortcut opens in a specific location and the notebook
directory may not be reachable from there.

Also added the license file.